requirements.txt
Flask==1
Pillow
matplotlib
requests
